bring back online claims as a basis
We looked into online claim chains again since that
is what we can use most easily with the current
implementation.

In fact a lot of the discussions so far have been
optimisation ideas to enable the inband usage.

Now we will first aim to get something working
with an online storage infrastructure and then see if
we need to move to inband claims.