gossip: cleanup
remove references to the calculation table.

The table was confusing and is hard to add to the document.
We provide some results in the text and that should be enough.

The calculations are of limited value anyway due to the
relevance of more targeted attacks. The numbers in the text
suffice for illustrating that consistent attacks are much harder.