group: metadata in vc protocol
No need to specify a separate key verification process for the verified group.
We can reuse the one we already have and just include a bit of metadata.

Since we currently only use the group name
to signal what Alice is about to do
there is no need to bounce that info back and force
In particular we do not need to protect
against Bob altering the group name
if Alice just stores it herself.

There was no evidence in the protocol
thus far
that Bob actually agreed
to become a member of the group.
This is still the case.

If we wanted to change it Bob would have
to send a message to the group to confirm
he wants to join.
A signature or so will not do
because the group members just learned Bobs key
through the same mechanism.