initial sequence diagram for verified contact flow
references #29.

Kept the sequence diagram close to the current numbering in the new.rst doc.
Would probably good to have different numbers like 3.a) 3.b)
for the separate parts of all the steps.

Also not using the resulting diagram in the pdf yet.
Diagram is fairly big and will probably result in a full page.
This should be fine as it allows comparing the instructions
and the diagram side by side.